Add build-net5.cake for .NET Core 3+ solutions
This script will eventually replace build-netcoreapp.cake as a standard Cake
template file for use in our builds.  The advantages for this build.cake
template is the addition of spinning up Docker images for SQL Server
and Elasticsearch as needed.  There is also an Azure emulator (Azurite)
container that will be added later.

The Docker containers are only created if the appropriate environment
variables are set prior to having Cake process the build.cake file.